refactor: remove barrel files and use direct imports
Replace index.ts re-export files with direct imports from source modules.
This improves code clarity, reduces circular dependency risks, and follows
module resolution best practices. All 35 files have been updated to import
directly from their source locations.

- Delete barrel files: src/components/charts/index.ts, src/components/overpay/index.ts, src/context/index.ts, src/hooks/index.ts, src/lib/loans/index.ts, src/types/index.ts
- Update imports across the codebase to use direct module paths
- Document preference against barrel files in CLAUDE.md
- All tests pass (164 tests)
