+++ src/app/overpay/page.tsx
@@ -1,6 +1,6 @@
 import { AppErrorBoundary } from "@/components/ErrorBoundary";
-import { OverpayPage } from "@/components/overpay";
-import { LoanProvider } from "@/context";
+import { OverpayPage } from "@/components/overpay/OverpayPage";
+import { LoanProvider } from "@/context/LoanContext";
 
 export default function OverpayRoute() {
   return (
