+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import App from "@/components/App";
 import { AppErrorBoundary } from "@/components/ErrorBoundary";
-import { LoanProvider } from "@/context";
+import { LoanProvider } from "@/context/LoanContext";
 
 export default function Home() {
   return (
