+++ src/components/ConfigPanel.tsx
@@ -9,7 +9,7 @@ import {
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Label } from "@/components/ui/label";
 import { Switch } from "@/components/ui/switch";
-import { useLoanContext } from "@/context";
+import { useLoanContext } from "@/context/LoanContext";
 
 export function ConfigPanel() {
   const { state, updateField } = useLoanContext();
