+++ src/components/CurrencyInput.tsx
@@ -1,5 +1,5 @@
 import { NumericFormat } from "react-number-format";
-import type { CurrencyInputProps } from "../types";
+import type { CurrencyInputProps } from "@/types/input";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
