--- src/components/charts/index.ts
@@ -1,6 +1,0 @@
-export {
-  ChartBase,
-  type ChartBaseProps,
-  type ChartSeriesConfig,
-  type ChartAnnotationConfig,
-} from "./ChartBase";
