--- src/components/overpay/index.ts
@@ -1,5 +1,0 @@
-export { OverpayInputs } from "./OverpayInputs";
-export { OverpayComparisonChart } from "./OverpayComparisonChart";
-export { OverpayVerdict } from "./OverpayVerdict";
-export { OverpaySummaryCards } from "./OverpaySummaryCards";
-export { OverpayPage } from "./OverpayPage";
