--- src/context/index.ts
@@ -1,9 +1,0 @@
-export { LoanProvider, useLoanContext } from "./LoanContext";
-export {
-  loanReducer,
-  initialState,
-  updateFieldAction,
-  resetAction,
-  applyPresetAction,
-  type LoanAction,
-} from "./loanReducer";
