--- src/hooks/index.ts
@@ -1,11 +1,0 @@
-export {
-  useLoanConfig,
-  useCurrentSalary,
-  useOverpayConfig,
-} from "./useStoreSelectors";
-export {
-  useTotalRepaymentData,
-  useRepaymentYearsData,
-  useInterestRateData,
-} from "./useChartData";
-export { useOverpayAnalysis } from "./useOverpayAnalysis";
