+++ src/hooks/useOverpayAnalysis.ts
@@ -3,8 +3,8 @@ import {
   useCurrentSalary,
   useOverpayConfig,
 } from "./useStoreSelectors";
-import type { OverpayAnalysisResult } from "@/lib/loans";
-import { simulateOverpayScenarios } from "@/lib/loans";
+import type { OverpayAnalysisResult } from "@/lib/loans/overpay-types";
+import { simulateOverpayScenarios } from "@/lib/loans/overpay-simulate";
 
 /**
  * Hook that performs overpay analysis calculations.
