--- src/types/index.ts
@@ -1,22 +1,0 @@
-// Store types
-export type { LoanState, LoanActions, LoanStore } from "./store";
-
-// Chart types
-export type { DataPoint } from "./chart";
-
-// Input component types
-export type {
-  CurrencyInputProps,
-  PercentageInputProps,
-  DateInputProps,
-  NumericFormatInputProps,
-} from "./input";
-
-// Loan calculation types (re-exported from lib/loans)
-export type {
-  PlanType,
-  UndergraduatePlanType,
-  Loan,
-  SimulationResult,
-  SimulationMapper,
-} from "@/lib/loans";
